requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
boto3==1.34.69
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.1
